% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createedges.R
\name{createedges}
\alias{createedges}
\title{Creates an untargeted graph from a biological sequence}
\usage{
createedges(sequence, word = 3, step = 1)
}
\arguments{
\item{sequence}{It is a vector that represents the sequence}

\item{word}{This integer parameter decides the size of the word that will
be formed}

\item{step}{It is the integer parameter that decides the step that will
be taken to make a new connection}
}
\value{
Returns the array used to creates the edge list
}
\description{
A function that from a biological sequence generates a graph
not addressed having as words vertices, this being able to have its size
parameter set by the' word 'parameter. The connections between words depend
of the' step 'parameter that indicates the next connection to be formed
}
\author{
Murilo Montanini Breve
}
