% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatingDataframe.R
\name{creatingDataframe}
\alias{creatingDataframe}
\title{Creates a feature matrix using complex network topological measures}
\usage{
creatingDataframe(measures, tamM, tamLNC, tamSNC)
}
\arguments{
\item{measures}{The complex network topological measures}

\item{tamM}{mRNA sequence size}

\item{tamLNC}{lncRNA sequence size}

\item{tamSNC}{snRNA sequence size}
}
\value{
Returns the feature matrix in scale 0-1
}
\description{
A function that from the complex network topological measures
create the feature matrix.
}
\author{
Murilo Montanini Breve
}
