% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveofentropy.R
\name{curveofentropy}
\alias{curveofentropy}
\title{Creates an entropy curve}
\usage{
curveofentropy(H, threshold)
}
\arguments{
\item{H}{The 'training' return for the entropy measures}

\item{threshold}{The 'training' return for the threshold}
}
\value{
Returns a entropy curve
}
\description{
A function that from the entropy measures and threshold creates
an entropy curve.
}
\author{
Murilo Montanini Breve
}
