% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Calculates the entropy}
\usage{
entropy(x)
}
\arguments{
\item{x}{The probabilities P0 and P1}
}
\value{
Returns the entropy
}
\description{
A function that calculates the entropy
}
\author{
Murilo Montanini Breve
}
