% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filtering}
\alias{filtering}
\title{Filters the edges}
\usage{
filtering(edgestoselect, edgestofilter)
}
\arguments{
\item{edgestoselect}{The selected edges}

\item{edgestofilter}{The edges used to filter}
}
\value{
Returns the filtered edges
}
\description{
A function that filters the edges after the maximum entropy is
obtained
}
\author{
Murilo Montanini Breve
}
