% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixmultiplication.R
\name{matrixmultiplication}
\alias{matrixmultiplication}
\title{Compares the matrices}
\usage{
matrixmultiplication(data, histodata)
}
\arguments{
\item{data}{Adjacency matrix}

\item{histodata}{'trainingResult' data}
}
\value{
Returns the filtered adjacency matrix
}
\description{
A function that compares the matrices 'trainingResult' and
the adjacency matrix to produce a filtered adjacency matrix.
}
\author{
Murilo Montanini Breve
}
