% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxentropy.R
\name{maxentropy}
\alias{maxentropy}
\title{Calculates the maximum entropy}
\usage{
maxentropy(histogram)
}
\arguments{
\item{histogram}{The histogram (used in 'training' function)}
}
\value{
Returns the maximum entropy
}
\description{
A function that calculates the maximum entropy
}
\author{
Murilo Montanini Breve
}
