% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Rescales the results between values from 0 to 1}
\usage{
preprocessing(datah, tamM, tamLNC, tamSNC)
}
\arguments{
\item{datah}{Array with results numerics}

\item{tamM}{Integer number of mRNA sequences}

\item{tamLNC}{Integer number of lncRNA sequences}

\item{tamSNC}{Integer number of sncRNA sequences}
}
\value{
Returns the array with the rescaled values
}
\description{
Given the results the data is rescaled for values between
0 and 1, so that the length of the sequences does not influence the results.
The rescaling of the sequences are made separately
}
\author{
Murilo Montanini Breve
}
