% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectingEdges.R
\name{selectingEdges}
\alias{selectingEdges}
\title{Selects the edges of the adjacency matrix}
\usage{
selectingEdges(MAX, data)
}
\arguments{
\item{MAX}{The maximum entropy}

\item{data}{The adjacency matrix}
}
\value{
Returns the selected edges of the adjacency matrix
}
\description{
A function that selects the edges of the adjacency matrix
}
\author{
Murilo Montanini Breve
}
