% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{training}
\alias{training}
\title{Trains the algorithm to select the edges that maximize the entropy}
\usage{
training(mRNA, lncRNA, sncRNA = NULL)
}
\arguments{
\item{mRNA}{Directory where the file .FASTA lies with the mRNA sequences}

\item{lncRNA}{Directory where the file .FASTA lies with the lncRNA sequences}

\item{sncRNA}{Directory where the file .FASTA lies with the sncRNA sequences
(optional)}
}
\value{
Returns the edge lists and the 'curveofentropy' function inputs
}
\description{
A function that trains the algorithm to select the edges that
maximize the entropy
}
\author{
Murilo Montanini Breve
}
